// src/types/statusCode.ts
export enum EStatusCodes {
  SUCCESS = 200,
  INVALID_USERNAME_OR_PASSWORD = 303,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  API_NOT_FOUND = 404,
  OLD_PASSWORD_NOT_CORRECT = 405,
  EMAIL_EXIST = 406,
  ITEM_EXIST = 407,
  ITEM_NOT_EXIST = 408,
  UPLOAD_ERROR = 409,
  IMPORT_CSV_FAIL = 410,
  ITEM_IN_USE = 411,
  WORK_ORDER_IS_ACTIVE = 412,
  PROJECT_IS_ACTIVE = 413,
  CHANGE_STATUS_NOT_SUCCESS = 414,
  LOCKED_ACCOUNT_TEMPORARILY = 415,
  LOCKED_ACCOUNT_PERMANENTLY = 416,
  INVALID_PARAMETER = 422,
  SYSTEM_ERROR = 500,
  SERVER_TEMPORARILY_UNAVAILABLE = 503,
}